game elements v1

+++ backend/src/main/java/pl/myc22ka/mathapp/user/dto/UserDTO.java
@@ -14,7 +14,7 @@ public record UserDTO(
         String lastname,
         String email,
         String role,
-        Integer points,
+        Double points,
         Integer level,
         String phoneNumber,
         String address,
